requirements.txt
dingtalk-stream>=0.24.0
python-dotenv>=0.19.0
